// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface HomeHero {
  id: number;
  tag?: string | null;
  titlePrefix?: string | null;
  titleEmphasis?: string | null;
  subtitle?: string | null;
  stats?:
    | {
        value?: string | null;
        label?: string | null;
        id?: string | null;
      }[]
    | null;
  memoriasTag?: string | null;
  memoriasTitle?: string | null;
  memoriasLead?: string | null;
  memoriasParagraphs?:
    | {
        text?: string | null;
        id?: string | null;
      }[]
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}